Reinstate dependency between UserStats and Stats. Remove references to UserStats in State (causing duplicate calls to UserStats's constructor).
